requirements.txt
# Core Web Framework
fastapi==0.135.2
uvicorn[standard]==0.42.0

# WebSocket Support
websockets==16.0
python-multipart==0.0.22

# Data Validation
pydantic==2.12.5
pydantic-settings==2.13.1

# Development Tools
python-dotenv==1.2.2

# Audio Transcoding
audioop-lts==0.2.1

# Google Cloud & AI
google-genai==1.68.0
